requirements.txt
pdf2image==1.17.0
Pillow==11.1.0
PyPDF2==3.0.1
streamlit==1.41.0
